requirements.txt

Django==3.2.14
django-admin-rangefilter==0.8.5
django-cors-headers==3.10.1
django-import-export==2.7.1
django-quill-editor==0.1.40
django-resized==1.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
redis==2.10.6
drf-yasg==1.20.0
celery==3.1.18
Pillow==8.4.0
PyJWT==1.7.1
channels==3.0.4
channels-redis==3.3.1
mysqlclient==2.1.0
